+++ pkg/kubelet/kuberuntime/BUILD
@@ -84,7 +84,7 @@ go_library(
             "//vendor/github.com/opencontainers/runc/libcontainer/cgroups/fs:go_default_library",
         ],
         "@io_bazel_rules_go//go/platform:windows": [
-            "//pkg/kubelet/apis:go_default_library",
+            "//staging/src/k8s.io/kubelet/pkg/apis:go_default_library",
         ],
         "//conditions:default": [],
     }),
