+++ staging/src/k8s.io/kubelet/BUILD
@@ -10,11 +10,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//staging/src/k8s.io/kubelet/config/v1beta1:all-srcs",
-        "//staging/src/k8s.io/kubelet/pkg/apis/deviceplugin/v1alpha:all-srcs",
-        "//staging/src/k8s.io/kubelet/pkg/apis/deviceplugin/v1beta1:all-srcs",
-        "//staging/src/k8s.io/kubelet/pkg/apis/pluginregistration/v1:all-srcs",
-        "//staging/src/k8s.io/kubelet/pkg/apis/pluginregistration/v1alpha1:all-srcs",
-        "//staging/src/k8s.io/kubelet/pkg/apis/pluginregistration/v1beta1:all-srcs",
+        "//staging/src/k8s.io/kubelet/pkg/apis:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
